feat(content): ship a buildable radar; fix an unbounded fog leak and an int32 overflow
Three things, two of them found by an independent reviewer rather than by me.

1. The radar existed as a mechanic with nothing to build.

Building.bIsRadar drives both the RadarDetected sweep in SystemFogOfWar and ADR-0013's
"Radar / minimap" row, and no shipped entity definition set it. Every radar test authored its
own definition at runtime, so the entire feature was reachable only from the test suite -- a
real match had no radar to construct.

Both playable factions now get a Radar Complex: 2x2, 1000 credits, 15 s, 75 power, T1,
prerequisite the power plant, produced by the construction yard. Localisation keys for it
already existed (radar_su / radar_al), which is its own small piece of evidence that this was
meant to exist and was simply never wired up.

Two deliberate choices, both load-bearing:
  - Power priority is left to DefaultPowerPriorityFor, which sees bIsRadar and assigns
    Auxiliary. The radar is what a deficit is *supposed* to take first; hardcoding the band
    here would let the content and the derivation disagree.
  - T1, not T2. From Severe, T2+ production is throttled, so a T2 radar destroyed during a
    blackout could not be rebuilt -- a power crisis would become permanent.
  - 75 power is deliberately expensive to run: the minimap going dark under load is the point
    of the ADR-0013 row, and a radar drawing almost nothing would never trigger it.

Two tests assert the shipped content rather than a hand-made definition, so deleting the
definition or moving it to T2 now breaks a test instead of silently reverting the feature.

2. DirtyRegions grew without bound (pre-existing, worsened by the radar).

DirtyRegions is a producer/consumer list for texture uploads: every reveal appends a rect and
the consumer is expected to drain it. Nothing in the shipping path ever called
ClearDirtyRegions -- only tests did. Measured in a real match: 2400 rects after 600 ticks with
three buildings, roughly 144k rects per player over a half-hour, for a list no shipping code
reads. RevealCircularArea had this from the start; the radar sweep added a second producer and
doubled the rate.

Now drained at the top of SystemFogOfWar, alongside ClearCurrentVisibility. Top rather than
bottom on purpose: a presentation layer that starts consuming this later must see the rects
produced by the tick that just ran.

3. Signed overflow in both reveal functions.

Radius * Radius and the per-cell distance were int32. Above a radius of 46340 that overflows,
which is undefined behaviour, not a wrong number. Both now compute in int64. No shipping
caller comes close today -- vision is derived from a content range and the radar sweep is a
fixed 24 -- but vision ranges are data, and data must not be able to make a function undefined.

Demonstrated with UBSan on the identical arithmetic, because the headless suite cannot catch
it: with int32, "signed integer overflow: 50000 * 50000 cannot be represented in type 'int'"
and the centre cell came out NeverSeen; with int64, no diagnostic and the centre visible. The
accompanying test pins the correct behaviour but is explicitly documented as *not* proving the
fix necessary on this platform -- arm64 -O2 happens to produce the same visible answer with
int32 restored, so the test still passes without the change. Saying so in the test is the
alternative to letting it read as coverage it does not provide.

Verified: cmake --build build/hb -j8   => clean, zero errors, zero warnings
          RA4Tests             => 521 passed, 0 failed
          RA4AITests           => 132 passed, 0 failed
          RA4PresentationTests =>  50 passed, 0 failed
          RA4InputTests        =>  67 passed, 0 failed
          Build.sh RedAlert4Editor Mac Development => Succeeded
Revert-verified: removing the radar's bIsRadar flag, or moving it to T2, or removing the
ClearDirtyRegions call each makes its own test fail. The overflow fix is the one change here
whose test cannot fail without it, and that is stated in the test rather than implied away.
NOT verified: no radar has been built in a running game -- Unreal was not launched.

+++ Source/RA4FogOfWar/Private/FogOfWarGrid.cpp
@@ -32,7 +32,11 @@ void FFogOfWarGrid::RevealCircularArea(int32_t PlayerIndex, int32_t CenterX, int
     const int32_t MinY = std::max(0, CenterY - Radius);
     const int32_t MaxY = std::min(Height - 1, CenterY + Radius);
 
-    const int32_t RadiusSq = Radius * Radius;
+    // int64: Radius * Radius overflows int32 above 46340, and signed overflow is undefined
+    // behaviour rather than merely a wrong answer. No caller passes anything near that today
+    // (vision is derived from a content range, radar is a fixed 24), but a map or a definition
+    // is data, and data must not be able to make this function undefined.
+    const int64_t RadiusSq = int64_t(Radius) * int64_t(Radius);
     bool bChanged = false;
 
     auto& Data = VisibilityData[PIdx];
@@ -41,7 +45,8 @@ void FFogOfWarGrid::RevealCircularArea(int32_t PlayerIndex, int32_t CenterX, int
     {
         for (int32_t X = MinX; X <= MaxX; ++X)
         {
-            const int32_t DistSq = (X - CenterX) * (X - CenterX) + (Y - CenterY) * (Y - CenterY);
+            const int64_t DistSq = int64_t(X - CenterX) * int64_t(X - CenterX) +
+                                   int64_t(Y - CenterY) * int64_t(Y - CenterY);
             if (DistSq <= RadiusSq)
             {
                 const size_t Idx = static_cast<size_t>(Y) * static_cast<size_t>(Width) + static_cast<size_t>(X);
@@ -71,7 +76,11 @@ void FFogOfWarGrid::RevealRadarArea(int32_t PlayerIndex, int32_t CenterX, int32_
     const int32_t MinY = std::max(0, CenterY - Radius);
     const int32_t MaxY = std::min(Height - 1, CenterY + Radius);
 
-    const int32_t RadiusSq = Radius * Radius;
+    // int64: Radius * Radius overflows int32 above 46340, and signed overflow is undefined
+    // behaviour rather than merely a wrong answer. No caller passes anything near that today
+    // (vision is derived from a content range, radar is a fixed 24), but a map or a definition
+    // is data, and data must not be able to make this function undefined.
+    const int64_t RadiusSq = int64_t(Radius) * int64_t(Radius);
     bool bChanged = false;
 
     auto& Data = VisibilityData[PIdx];
@@ -80,7 +89,8 @@ void FFogOfWarGrid::RevealRadarArea(int32_t PlayerIndex, int32_t CenterX, int32_
     {
         for (int32_t X = MinX; X <= MaxX; ++X)
         {
-            const int32_t DistSq = (X - CenterX) * (X - CenterX) + (Y - CenterY) * (Y - CenterY);
+            const int64_t DistSq = int64_t(X - CenterX) * int64_t(X - CenterX) +
+                                   int64_t(Y - CenterY) * int64_t(Y - CenterY);
             if (DistSq > RadiusSq)
             {
                 continue;
